requirements.txt
requests>=2.31.0
supabase>=2.3.0
python-dotenv>=1.0.0
playwright>=1.44
pandas>=2.2
openpyxl>=3.1
